# [RouteNode] Combines the prev reference into children 