In order to reduce the cost of the `RouteNode` object, this change stores the `prev` reference at the 0th index of the `children` array.  This change does increase route time ~7%:

Baseline:
```
Route design:                          302.50s
?? Initialization:                      14.05s
?  ?? determine route targets:          13.08s
?? Routing:                            288.44s
   ?? route clock:                       0.00s
   ?? route static nets:                 0.00s
   ?? route wire nets:                 276.56s
   ?  ?? rnodes creation:               43.62s
   ?  ?? route connections:            212.65s
   ?  ?? update timing:                  0.00s
   ?  ?? update congestion costs:       20.29s
   ?? finalize routes:                  11.88s
```

With this PR:
```
Route design:                          325.20s
?? Initialization:                      14.31s
?  ?? determine route targets:          13.28s
?? Routing:                            310.88s
   ?? route clock:                       0.00s
   ?? route static nets:                 0.00s
   ?? route wire nets:                 299.34s
   ?  ?? rnodes creation:               62.82s
   ?  ?? route connections:            215.09s
   ?  ?? update timing:                  0.00s
   ?  ?? update congestion costs:       21.44s
   ?? finalize routes:                  11.54s
```